Move tests and enable them
